+++ packages/ui-components/web/src/Icons/index.ts
@@ -20,6 +20,7 @@ export { default as TrashIcon } from './TrashIcon';
 export { default as SoundCloudIcon } from './SoundCloudIcon';
 export { default as SpotifyIcon } from './SpotifyIcon';
 export { default as LoaderIcon } from './LoaderIcon';
+export { default as CirclePlusIcon } from './CirclePlusIcon';
 
 //need to import from plugin-commons
 // export { default as AlignCenterIcon } from './AlignCenterIcon';